Refactoring row update
https://github.com/metasfresh/metasfresh/issues/6949
