+++ backend/metasfresh-webui-api/src/main/java/de/metas/ui/web/pickingV2/productsToPick/rows/ProductsToPickRowsService.java
@@ -19,7 +19,6 @@ import de.metas.i18n.ITranslatableString;
 import de.metas.ui.web.pickingV2.config.PickingConfigRepositoryV2;
 import de.metas.ui.web.pickingV2.config.PickingConfigV2;
 import de.metas.ui.web.pickingV2.packageable.PackageableRow;
-import de.metas.ui.web.pickingV2.productsToPick.process.WebuiPickHUResult;
 import de.metas.ui.web.pickingV2.productsToPick.rows.factory.ProductsToPickRowsDataFactory;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.ui.web.window.model.lookup.LookupDataSourceFactory;
